Refactor diff parsing and file details out of viewer state
