+++ web/src/lib/components/diff-filtering/DiffFilterDialog.svelte
@@ -1,7 +1,8 @@
 <script lang="ts">
+    import { FILE_STATUSES } from "$lib/file-details";
     import { getFileStatusProps } from "$lib/diff-viewer.svelte";
     import { Button, Dialog, ToggleGroup } from "bits-ui";
-    import { tryCompileRegex, FILE_STATUSES } from "$lib/util";
+    import { tryCompileRegex } from "$lib/util";
     import { slide } from "svelte/transition";
     import { type DiffFilterDialogProps, type FilePathFilterMode } from "$lib/components/diff-filtering/index.svelte";
     import { GlobalOptions } from "$lib/global-options.svelte";
