+++ web/src/lib/components/sidebar/Sidebar.svelte
@@ -1,5 +1,7 @@
 <script lang="ts">
-    import { type FileDetails, getFileStatusProps, MultiFileDiffViewerState, staticSidebar } from "$lib/diff-viewer.svelte";
+    import { type FileDetails } from "$lib/file-details";
+
+    import { getFileStatusProps, MultiFileDiffViewerState, staticSidebar } from "$lib/diff-viewer.svelte";
     import Tree from "$lib/components/tree/Tree.svelte";
     import { type TreeNode } from "$lib/components/tree/index.svelte";
     import { on } from "svelte/events";
