+++ web/src/lib/util.ts
@@ -1,4 +1,3 @@
-import { type FileDetails, type ImageFileDetails, LoadingState, makeTextDetails } from "./diff-viewer.svelte";
 import type { BundledLanguage, SpecialLanguage } from "shiki";
 import { onMount } from "svelte";
 import { on } from "svelte/events";
@@ -83,112 +82,6 @@ export async function bytesEqual(
     }
 
     return true;
-}
-
-export type FileStatus = "added" | "removed" | "modified" | "renamed" | "renamed_modified";
-export const FILE_STATUSES: FileStatus[] = ["added", "removed", "modified", "renamed", "renamed_modified"];
-
-export function binaryFileDummyDetails(fromFile: string, toFile: string, status: FileStatus): FileDetails {
-    let fakeContent: string;
-    switch (status) {
-        case "added":
-            fakeContent = `diff --git a/${toFile} b/${toFile}\n--- /dev/null\n+++ b/${toFile}\n@@ -0,0 +1,1 @@\n+Cannot show binary file`;
-            break;
-        case "removed":
-            fakeContent = `diff --git a/${fromFile} b/${fromFile}\n--- a/${fromFile}\n+++ /dev/null\n@@ -1,1 +0,0 @@\n-Cannot show binary file`;
-            break;
-        default:
-            fakeContent = `diff --git a/${fromFile} b/${toFile}\n--- a/${fromFile}\n+++ b/${toFile}\n@@ -1,1 +1,1 @@\n-Cannot show binary file\n+Cannot show binary file`;
-            break;
-    }
-    return makeTextDetails(fromFile, toFile, status, fakeContent);
-}
-
-const fileRegex = /diff --git a\/(\S+) b\/(\S+)\r?\n(?:.+\r?\n)*?(?=-- \r?\n|diff --git|$)/g;
-
-type BasicHeader = {
-    fromFile: string;
-    toFile: string;
-    status: FileStatus;
-    binary: boolean;
-};
-
-function parseHeader(patch: string, fromFile: string, toFile: string): BasicHeader {
-    let status: FileStatus = "modified";
-    if (fromFile !== toFile) {
-        status = "renamed_modified";
-    }
-    let binary = false;
-    let foundIndex = false;
-
-    let lineStart = 0;
-    while (true) {
-        const lineEnd = patch.indexOf("\n", lineStart);
-        if (lineEnd === -1) {
-            break; // No more lines
-        }
-        const line = patch.substring(lineStart, lineEnd);
-        if (line.startsWith("similarity index 100%")) {
-            status = "renamed";
-            if (isImageFile(fromFile) && isImageFile(toFile)) {
-                binary = true; // Treat renamed images as binary
-            }
-        } else if (line.startsWith("deleted file mode")) {
-            status = "removed";
-        } else if (line.startsWith("new file mode")) {
-            status = "added";
-        } else if (line.startsWith("index ")) {
-            foundIndex = true;
-        } else if (foundIndex) {
-            if (line.startsWith("Binary files")) {
-                binary = true;
-            }
-            // end of header
-            break;
-        }
-        lineStart = lineEnd + 1;
-    }
-
-    return { fromFile, toFile, status, binary };
-}
-
-export function parseMultiFilePatch(
-    patchContent: string,
-    loadingState: LoadingState,
-    imageFactory?: (fromFile: string, toFile: string, status: FileStatus) => ImageFileDetails | null,
-): AsyncGenerator<FileDetails> {
-    const split = splitMultiFilePatch(patchContent);
-    loadingState.totalCount = split.length;
-    async function* detailsGenerator() {
-        for (const [header, content] of split) {
-            if (header.binary) {
-                if (imageFactory !== undefined && isImageFile(header.fromFile) && isImageFile(header.toFile)) {
-                    const imageDetails = imageFactory(header.fromFile, header.toFile, header.status);
-                    if (imageDetails != null) {
-                        yield imageDetails;
-                        continue;
-                    }
-                } else {
-                    yield binaryFileDummyDetails(header.fromFile, header.toFile, header.status);
-                    continue;
-                }
-            }
-
-            yield makeTextDetails(header.fromFile, header.toFile, header.status, content);
-        }
-    }
-    return detailsGenerator();
-}
-
-export function splitMultiFilePatch(patchContent: string): [BasicHeader, string][] {
-    const patches: [BasicHeader, string][] = [];
-    let fileMatch;
-    while ((fileMatch = fileRegex.exec(patchContent)) !== null) {
-        const [fullFileMatch, fromFile, toFile] = fileMatch;
-        const header = parseHeader(fullFileMatch, fromFile, toFile);
-        patches.push([header, fullFileMatch]);
-    }
-    return patches;
 }
 
 const imageExtensions: Set<string> = new Set(["jpg", "jpeg", "png", "gif", "webp", "bmp", /*"svg",*/ "tiff", "ico"]);
